refactor: Remove @solana/wallet-adapter-wallets re-exports dep
