+++ package.json
@@ -63,7 +63,8 @@
     "@changesets/changelog-git": "^0.2.0",
     "@changesets/cli": "^2.27.1",
     "@ianvs/prettier-plugin-sort-imports": "^4.2.1",
-    "@solana/wallet-adapter-wallets": "^0.19.32",
+    "@solana/wallet-adapter-coinbase": "^0.1.19",
+    "@solana/wallet-adapter-trust": "^0.1.13",
     "@tanstack/react-query": "^5.28.4",
     "@testing-library/react": "^14.2.2",
     "@types/lodash": "^4.14.202",
