+++ packages/solana/src/solana-provider/__tests__/switch-chain.test.tsx
@@ -3,8 +3,8 @@
 import { ConnectButton, Connector, useProvider } from '@ant-design/web3';
 import { metadata_CoinbaseWallet } from '@ant-design/web3-assets';
 import { SolanaDevnet } from '@ant-design/web3-assets/solana';
+import { CoinbaseWalletAdapter } from '@solana/wallet-adapter-coinbase';
 import type { ConnectionContextState } from '@solana/wallet-adapter-react';
-import { CoinbaseWalletAdapter } from '@solana/wallet-adapter-wallets';
 import { fireEvent } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
